fix: backend compilation
